Add a PowerShell build script.
This change adds a flexible build script written in PowerShell. My team
and I have found this useful because we are only interested in building
one of the products for a single architecture, and we aren't building
on Linux. This script is flixible and cross platform to Linux, Windows,
and Mac.

For usage instructions you can read the help text in the file, for at
the command line you can invoke Get-Help ./build.ps1 -full.